Explicitly define the namespace in manifests
This is needed in order to render the templates with the correct
namespace when passed the --namespace flag. This is required for tools
like Spinnaker.

This doesn't affect installing directly from helm since helm adds the
value directly from the --namespace flag.

For reference https://github.com/helm/helm/issues/5465